Rename mainContent to Blocks
